Added specific headers for new objects and utils.h
