+++ CMakeLists.txt
@@ -12,7 +12,7 @@ set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH}
 	"${CMAKE_SOURCE_DIR}/cmake/"
 )
 
-file(GLOB_RECURSE SRC_FILES src/*.c)
+file(GLOB_RECURSE SRC_FILES "src/*.c" "src/objects/*.c")
 add_compile_options(
 	-Wall
 	-Wextra
@@ -26,7 +26,8 @@ add_executable(wtf
 
 target_include_directories(wtf
 	PRIVATE
-	${PROJECT_SOURCE_DIR}/include
+	"${PROJECT_SOURCE_DIR}/include"
+	"${PROJECT_SOURCE_DIR}/include/objects"
 )
 
 find_package(SDL2 REQUIRED)
