+++ include/utils.h
@@ -1,0 +1,10 @@
+#ifndef WHEN_THE_FACTORY_UTILS_
+#define WHEN_THE_FACTORY_UTILS_
+
+/* Coords of a tile on the map. */
+struct TileCoords {
+	int x, y;
+};
+typedef struct TileCoords TileCoords;
+
+#endif
